Replace all invalid characters in local identifier
The `buildTag` used in the local identifier passed to BrowserStack
can contain invalid characters, particularly if the job is in a
folder. The current implementation replaces specific characters,
which misses all other invalid ones.

This change replaces any characters which don't match the set of
valid identifiers, obtained from the error message which prompted
the bug report:

> identifier can only have alphanumeric characthers with combination of hiphen(-), underscore(_) and dot(.)